# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BROTE Communication Platform API
  description: API de BCP — plataforma de comunicación empresarial omnicanal
  version: 0.1.0
  contact:
    name: BROTE
servers:
  - url: http://localhost:3000
    description: Desarrollo local
paths:
...
  /auth/register:
    post:
      summary: Registrar workspace y usuario dueño
      description: 'Password must meet complexity requirements: minimum 8 characters, 1 uppercase, 1 lowercase, 1 number, 1 special character'
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Workspace y usuario creados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/AuthTokensWithIds' }
        '400':
          description: Error de validación (campos faltantes, inválidos, o business rules)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /auth/login:
    post:
      summary: Iniciar sesión
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Credenciales válidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/AuthTokens' }
        '400':
          description: Error de validación (email o password faltantes/inválidos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Email o password inválidos (credenciales rechazadas)
  /auth/refresh:
    post:
      summary: Renovar access token usando el refresh token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Tokens renovados
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/AuthTokens' }
        '400':
          description: Error de validación (refreshToken faltante o inválido)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ValidationError:
      type: object
      properties:
        field: { type: string, description: 'Campo con error' }
        message: { type: string, description: 'Mensaje de error específico' }
    ValidationErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: string
          description: 'Mensaje de error único (si solo hay un error)'
        errors:
          type: array
          description: 'Array de errores por campo (si hay múltiples errores)'
          items: { $ref: '#/components/schemas/ValidationError' }
      example:
        success: false
        error: 'ownerPassword must be at least 8 characters'
    RegisterRequest:
      type: object
      required: [ownerName, ownerEmail, ownerPassword, workspaceName]
      properties:
        ownerName: { type: string }
        ownerEmail: { type: string, format: email }
        ownerPassword:
          type: string
          minLength: 8
          description: 'Must contain: uppercase + lowercase + number + special character'
          example: 'MyPassword123!'
        workspaceName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, description: 'Must meet complexity requirements (uppercase + lowercase + number + special character)' }
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: 'Segundos hasta la expiración del access token' }
    AuthTokensWithIds:
      allOf:
        - $ref: '#/components/schemas/AuthTokens'
        - type: object
          properties:
            workspaceId: { type: string }
            userId: { type: string }
...